Add error handling and file management functions in process_edit.php; include new files_helps.php for directory checks and to_do functionality
