+++ .gitignore
@@ -11,3 +11,4 @@ publish_reports/_index.php
 /reports
 test.php
 tx.php
+bots/new_x.php
