+++ start.php
@@ -9,57 +9,10 @@ if ($_SERVER['REQUEST_METHOD'] !== 'POST') {
 use function Publish\Helps\pub_test_print;
 use function Publish\AccessHelps\get_access_from_db;
 use function Publish\AccessHelpsNew\get_access_from_db_new;
-use function Publish\WD\LinkToWikidata;
 use function Publish\TextFix\DoChangesToText;
 use function WpRefs\FixPage\DoChangesToText1;
 use function Publish\EditProcess\processEdit;
-
-// $rand_id = rand(0, 999999999);
-$rand_id = time() .  "-" . bin2hex(random_bytes(6));
-
-// $main_dir = check_dirs($rand_id, "reports");
-$main_dir_by_day = check_dirs($rand_id, "reports_by_day");
-
-function check_dirs($rand_id, $reports_dir_main)
-{
-    $publish_reports = "I:/mdwiki/publish-repo/publish_reports/";
-    // ---
-    if (!is_dir($publish_reports)) {
-        $publish_reports = __DIR__ . "/../publish_reports/";
-    }
-    // ---
-    $reports_dir = "$publish_reports/$reports_dir_main/";
-    // ---
-    if (!is_dir($reports_dir)) {
-        mkdir($reports_dir, 0755, true);
-    }
-    // ---
-    $year_dir = $reports_dir . date("Y");
-    // ---
-    if (!is_dir($year_dir)) {
-        mkdir($year_dir, 0755, true);
-    }
-    // ---
-    $month_dir = $year_dir . "/" . date("m");
-    // ---
-    if (!is_dir($month_dir)) {
-        mkdir($month_dir, 0755, true);
-    }
-    // ---
-    $day_dir = $month_dir . "/" . date("d");
-    // ---
-    if (!is_dir($day_dir)) {
-        mkdir($day_dir, 0755, true);
-    }
-    // ---
-    $main1_dir = $day_dir . "/" . $rand_id;
-    // ---
-    if (!is_dir($main1_dir)) {
-        mkdir($main1_dir, 0755, true);
-    }
-    // ---
-    return $main1_dir;
-}
+use function Publish\FilesHelps\to_do;
 
 function get_revid($sourcetitle)
 {
@@ -72,7 +25,7 @@ function get_revid($sourcetitle)
         $json = json_decode(file_get_contents($revids_file), true);
         $revid = $json[$sourcetitle] ?? "";
         return $revid;
-    } catch (Exception $e) {
+    } catch (\Exception $e) {
         pub_test_print($e->getMessage());
     }
     return "";
@@ -81,33 +34,6 @@ function get_revid($sourcetitle)
 function make_summary($revid, $sourcetitle, $to, $hashtag)
 {
     return "Created by translating the page [[:mdwiki:Special:Redirect/revision/$revid|$sourcetitle]] to:$to $hashtag";
-}
-
-function to_do($tab, $file_name)
-{
-    global $main_dir_by_day; // $main_dir,
-    // ---
-    $tab['time'] = time();
-    $tab['time_date'] = date("Y-m-d H:i:s");
-    // ---
-    /*
-    try {
-        // dump $tab to file in folder to_do
-        $file_j = $main_dir . "/$file_name.json";
-        // ---
-        file_put_contents($file_j, json_encode($tab, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE));
-    } catch (Exception $e) {
-        pub_test_print($e->getMessage());
-    }*/
-    // ---
-    try {
-        // dump $tab to file in folder to_do
-        $file_j = $main_dir_by_day . "/$file_name.json";
-        // ---
-        file_put_contents($file_j, json_encode($tab, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE));
-    } catch (Exception $e) {
-        pub_test_print($e->getMessage());
-    }
 }
 
 function formatTitle($title)
@@ -156,50 +82,6 @@ function handleNoAccess($user, $tab)
     // file_put_contents(__DIR__ . '/editit.json', json_encode($editit, JSON_PRETTY_PRINT));
 }
 
-function handleSuccessfulEdit($sourcetitle, $lang, $user, $title, $access_key, $access_secret)
-{
-    $LinkTowd = [];
-    // ---
-    try {
-        $LinkTowd = LinkToWikidata($sourcetitle, $lang, $user, $title, $access_key, $access_secret) ?? [];
-        // ---
-    } catch (Exception $e) {
-        pub_test_print($e->getMessage());
-    }
-    // ---
-    if (isset($LinkTowd['error'])) {
-        $tab3 = [
-            'error' => $LinkTowd['error'],
-            'qid' => $LinkTowd['qid'] ?? "",
-            'title' => $title,
-            'sourcetitle' => $sourcetitle,
-            'lang' => $lang,
-            'username' => $user
-        ];
-        // if str($LinkTowd['error']) has "Links to user pages"  then file_name='wd_user_pages' else 'wd_errors'
-        // ---
-        $file_name = 'wd_errors';
-        // ---
-        $errs = [
-            'Links to user pages' => "wd_user_pages",
-            'get_csrftoken' => "wd_csrftoken",
-            'protectedpage' => "wd_protectedpage",
-        ];
-        // ---
-        $f_text = json_encode($LinkTowd['error']);
-        // ---
-        foreach ($errs as $err => $file) {
-            if (strpos($f_text, $err) !== false) {
-                $file_name = $file;
-                break;
-            }
-        }
-        // ---
-        to_do($tab3, $file_name);
-    }
-    // ---
-    return $LinkTowd;
-}
 
 function start2($request, $user, $access, $tab)
 {
@@ -223,16 +105,14 @@ function start2($request, $user, $access, $tab)
         $text = $newtext;
     }
     // ---
-    $tabx = processEdit($request, $access, $text, $user, $tab);
+    $editit = processEdit($request, $access, $text, $user, $tab);
     // ---
     pub_test_print("\n<br>");
     pub_test_print("\n<br>");
     // ---
-    print(json_encode($tabx['editit'], JSON_PRETTY_PRINT));
+    print(json_encode($editit, JSON_PRETTY_PRINT));
     // ---
     // file_put_contents(__DIR__ . '/editit.json', json_encode($editit, JSON_PRETTY_PRINT));
-    // ---
-    to_do($tabx['tab'], $tabx['to_do_file']);
 }
